requirements.txt
cachetools==6.2.4
certifi==2026.1.4
charset-normalizer==3.4.4
cssselect==1.3.0
cssutils==2.11.1
idna==3.11
IMAPClient==3.1.0
lxml==6.0.2
more-itertools==10.8.0
premailer==3.10.0
python-dotenv==1.2.1
requests==2.32.5
urllib3==2.6.3
yagmail==0.15.293
